requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.1.1
cryptography==41.0.7

# Web framework and security
requests==2.31.0
urllib3==2.1.0
certifi==2023.11.17
fastapi>=0.68.0
uvicorn>=0.15.0
pillow>=8.3.1
python-multipart>=0.0.5

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1